requirements.txt
python-telegram-bot==21.4
apscheduler==3.10.4
tzdata==2025.1
